repos/at-exec-sim/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-json-logger==2.0.7
structlog==23.2.0

# NATS messaging
nats-py==2.6.0

# Prometheus metrics
prometheus-client==0.19.0

# Schema validation
jsonschema==4.20.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-benchmark==4.0.0
httpx==0.25.1
testcontainers==3.7.1